Se agregan test para el RepositorioMovimiento

+++ src/main/java/com/tallerwebi/infraestructura/RepositorioMovimientoImpl.java
@@ -4,6 +4,7 @@ import com.tallerwebi.dominio.excepcion.ExcepcionBaseDeDatos;
 import com.tallerwebi.dominio.excepcion.ExcepcionMovimientoNoEncontrado;
 import com.tallerwebi.dominio.movimiento.Movimiento;
 import com.tallerwebi.dominio.movimiento.RepositorioMovimiento;
+import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,7 +25,7 @@ public class RepositorioMovimientoImpl implements RepositorioMovimiento {
     }
 
     @Override
-    public List<Movimiento> obtenerMovimientos(Long idUsuario) throws ExcepcionBaseDeDatos{
+    public List<Movimiento> obtenerMovimientos(Long idUsuario) throws ExcepcionBaseDeDatos{ //BUSCA POR ID DE USUARIO
         try{
             Session session = sessionFactory.getCurrentSession();
             return session.createQuery("FROM Movimiento M WHERE M.usuario.id = :idUsuario", Movimiento.class)
@@ -36,19 +37,26 @@ public class RepositorioMovimientoImpl implements RepositorioMovimiento {
     }
 
     @Override
-    public Movimiento obtenerMovimientoPorId(Long id) throws ExcepcionMovimientoNoEncontrado {
-        Session session = sessionFactory.getCurrentSession();
-        Movimiento movimiento = session.createQuery("FROM Movimiento M WHERE M.id = :idMovimiento", Movimiento.class)
-                .setParameter("idMovimiento", id)
-                .uniqueResult();
-        if(movimiento==null){
-            throw new ExcepcionMovimientoNoEncontrado("No se encontro el movimiento");
+    public Movimiento obtenerMovimientoPorId(Long id) throws ExcepcionMovimientoNoEncontrado, ExcepcionBaseDeDatos { //BUSCA POR ID DE MOVIMIENTO
+        try {
+            Session session = sessionFactory.getCurrentSession();
+            Movimiento movimiento = session.createQuery("FROM Movimiento M WHERE M.id = :idMovimiento", Movimiento.class)
+                    .setParameter("idMovimiento", id)
+                    .uniqueResult();
+            if(movimiento==null)
+                throw new ExcepcionMovimientoNoEncontrado("No se encontro el movimiento");
+
+            return movimiento;
+        } catch (HibernateException he) {
+            throw new ExcepcionBaseDeDatos("Base de datos no disponible", he);
         }
-        return movimiento;
     }
 
     @Override
-    public void actualizarMovimiento(Movimiento movimiento) throws ExcepcionBaseDeDatos {
+    public void actualizarMovimiento(Movimiento movimiento) throws ExcepcionBaseDeDatos, ExcepcionMovimientoNoEncontrado {
+        if(movimiento==null)
+            throw new ExcepcionMovimientoNoEncontrado("No se encontro el movimiento");
+
         try{
             Session session = sessionFactory.getCurrentSession();
             session.update(movimiento);
@@ -68,17 +76,34 @@ public class RepositorioMovimientoImpl implements RepositorioMovimiento {
     }
 
     @Override
-    public void eliminarMovimiento(Movimiento movimiento) {
-        Session session = sessionFactory.getCurrentSession();
-        session.delete(movimiento);
+    public void eliminarMovimiento(Movimiento movimiento) throws ExcepcionBaseDeDatos, ExcepcionMovimientoNoEncontrado {
+        if (movimiento == null || movimiento.getId() == null)
+            throw new ExcepcionMovimientoNoEncontrado("No se encontro el movimiento");
+
+        try {
+            Session session = sessionFactory.getCurrentSession();
+            Movimiento movimientoExistente = session.get(Movimiento.class, movimiento.getId());
+            if (movimientoExistente == null)
+                throw new ExcepcionMovimientoNoEncontrado("No se encontro el movimiento");
+
+            session.delete(movimiento);
+        } catch (HibernateException he) {
+            throw new ExcepcionBaseDeDatos("Base de datos no disponible", he);
+        } catch (ExcepcionMovimientoNoEncontrado e) {
+            System.err.println("Error: " + e.getMessage());
+        }
     }
 
     @Override
-    public List<Movimiento> obtenerMovimientosPorFecha(Long idUsuario, LocalDate fecha) {
-        Session session = sessionFactory.getCurrentSession();
-        return session.createQuery("FROM Movimiento m WHERE m.usuario.id = :idUsuario AND m.fechayHora = :fecha", Movimiento.class)
-                .setParameter("idUsuario", idUsuario)
-                .setParameter("fecha", fecha)
-                .getResultList();
+    public List<Movimiento> obtenerMovimientosPorFecha(Long idUsuario, LocalDate fecha) throws ExcepcionBaseDeDatos {
+        try{
+            Session session = sessionFactory.getCurrentSession();
+            return session.createQuery("FROM Movimiento m WHERE m.usuario.id = :idUsuario AND m.fechayHora = :fecha", Movimiento.class)
+                    .setParameter("idUsuario", idUsuario)
+                    .setParameter("fecha", fecha)
+                    .getResultList();
+        }catch (HibernateException he) {
+            throw new ExcepcionBaseDeDatos("Base de datos no disponible", he);
+        }
     }
 }
